Change ORG_OPENCASTPROJECT_DOWNLOAD_URL in multiserver docker-compose
The Environment variable ORG_OPENCASTPROJECT_DOWNLOAD_URL sets org.opencastproject.download.url
which defaults to org.opencastproject.server.url. The setting matters only at the presentation
node and should match the url to the presentation node or a load balancer in front of it.
Since org.opencastproject.server.url is not reachable from the host in the samples it is
explicitly set to the presentation node URL reachable from the host.